+++ src/pages/Attendance/DayTable/index.jsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import SideBar from "../../components/SideBar";
+import SideBar from "../../../components/SideBar";
 import Table from "./Table"
 const Host = () => {
