+++ src/shared/Router.js
@@ -18,6 +18,7 @@ const Router = () => {
         <Route path="/classes/:id" element={PAGE.member} />
         <Route path="/host" element={PAGE.host} />
         <Route path="/attendance" element={PAGE.attendance} />
+        <Route path="/attendance2" element={PAGE.attendance2} />
         <Route
           path="/host/:classroomId/:scheduleId?/:timeId?"
           element={PAGE.host}
